sentiment/requirements.txt
requests
python-dotenv
mysql-connector-python
pandas
numpy
